requirements.txt
fastapi
loguru
Pillow
pipeit
pydantic
requests
uvicorn
onnxruntime
tqdm
funcaptcha_challenger
pytesseract